remove dublicated test class for azure shell
